Move static methods back to TeamCityExtension
