+++ src/main/groovy/org/labkey/gradle/task/RunTestSuite.groovy
@@ -18,6 +18,7 @@ package org.labkey.gradle.task
 import org.apache.commons.lang3.StringUtils
 import org.gradle.api.tasks.Internal
 import org.labkey.gradle.plugin.TeamCity
+import org.labkey.gradle.plugin.extension.TeamCityExtension
 import org.labkey.gradle.util.DatabaseProperties
 
 /**
@@ -49,7 +50,7 @@ abstract class RunTestSuite extends RunUiTest
 
     protected void setTeamCityProperties()
     {
-        if (TeamCityPropertiesTask.isOnTeamCity(project))
+        if (TeamCityExtension.isOnTeamCity(project))
         {
             systemProperty "teamcity.tests.recentlyFailedTests.file", project.teamcity['teamcity.tests.recentlyFailedTests.file']
             systemProperty "teamcity.build.changedFiles.file", project.teamcity['teamcity.build.changedFiles.file']
