+++ src/main/groovy/org/labkey/gradle/util/BuildUtils.groovy
@@ -33,7 +33,6 @@ import org.labkey.gradle.plugin.extension.LabKeyExtension
 import org.labkey.gradle.plugin.extension.ModuleExtension
 import org.labkey.gradle.plugin.extension.ServerDeployExtension
 import org.labkey.gradle.plugin.extension.TeamCityExtension
-import org.labkey.gradle.task.TeamCityPropertiesTask
 
 import java.nio.charset.StandardCharsets
 import java.nio.file.Files
@@ -262,7 +261,7 @@ class BuildUtils
         if (!project.projectDir.exists() && project.hasProperty("downloadLabKeyModules"))
             reasons.add("Project directory ${project.projectDir} does not exist.")
         String propValue = project.hasProperty(property) ? project.property(property) : null
-        String value = TeamCityPropertiesTask.getTeamCityProperty(project, property, propValue)
+        String value = TeamCityExtension.getTeamCityProperty(project, property, propValue)
         if (value == null)
         {
             reasons.add("Project does not have ${property} property")
@@ -492,8 +491,8 @@ class BuildUtils
     static Properties getStandardVCSProperties(project)
     {
         String buildNumber =
-                (String) TeamCityPropertiesTask.getTeamCityProperty(project, "system.teamcity.agent.dotnet.build_id", // Unique build ID
-                        TeamCityPropertiesTask.getTeamCityProperty(project,"build.number", null))
+                (String) TeamCityExtension.getTeamCityProperty(project, "system.teamcity.agent.dotnet.build_id", // Unique build ID
+                        TeamCityExtension.getTeamCityProperty(project,"build.number", null))
         Properties ret = new Properties()
         if (project.plugins.hasPlugin("org.labkey.versioning"))
         {
